Additional characters to InputfieldPageName.module
Added Turkish specific characters of "ığş" to the $defaultReplacements array, which were replaced with a dash. 

The uppercase "i" letter "İ", adds a dash after the character, e.g. "i-". Will solve this, but not right now. 

For further reference, in Turkish, uppercase "ı" is "I", and uppercase "i" is "İ". Turkish specific additions to Latin-5 are "çığöşüÇİĞÖŞÜ" and publicly used "âÂ".

As told in Wikipedia,
The Turkish alphabet consists of 29 letters (q, x, w omitted and ç, ş, ğ, ı, ö, ü added) 
The complete list is,
a, b, c, ç, d, e, f, g, ğ, h, ı, i, j, k, l, m, n, o, ö, p, r, s, ş, t, u, ü, v, y, and z (Note that capital of i is İ and lowercase I is ı.)